A few spelling or grammar issues
Great article!  I just noticed a few misspellings or missing words, or things like “an” when you meant “and” that are so darn easy to miss when proofreading and that spellcheckers miss too. A few it’s versus its issues (the bane of my existence).  Not sure I always got the it’s (it is/it has) versus its (possessive) right, so check MY proposed changes before believing me!   There is one sentence that ends in “for” which I did NOT change.  You might want to move that “for” earlier in the sentence.     Again, great article!  I edit/censor too much myself.    A lot of people who write books do so longhand first, precisely to (as you say) let things flow.  Then they type it up and edit it later. I’m a recent fountain pen fan and learned the tip about writing it out longhand first about several of my favorite writers such as Joe Haldeman, Neil Gaiman, etc.